User center - refactor the way of adding transparent nav bar
