+++ SwiftGG/Modules/UserCenter/ViewControllers/SGUserViewController.swift
@@ -86,7 +86,7 @@ extension SGUserViewController {
 }
 
 // MARK: - Helper
-extension SGUserViewController {
+extension SGUserViewController: TransparentNavBarProtocol {
     private func setupViews() {
         view.backgroundColor = UIColor(rgba: "#3595BF")
         tableView.registerNib(UINib(nibName: "SGUserReadingTableViewCell", bundle: nil), forCellReuseIdentifier: String(SGUserReadingTableViewCell))
@@ -94,25 +94,10 @@ extension SGUserViewController {
         tableView.delegate = self
         pinnedHeaderView.delegate = self
         
-        addCustomNavigationBar()
-    }
-
-    private func addCustomNavigationBar() {
-        let customNavigationBar = UINavigationBar(frame: CGRectMake(0, 0, CGRectGetWidth(self.navigationController!.navigationBar.bounds), 64))
-        customNavigationBar.tintColor = UIColor.whiteColor()
-        customNavigationBar.tintAdjustmentMode = .Normal
-        customNavigationBar.alpha = 1
+        let transparentNavBar = transparentNavigationBar()
+        view.addSubview(transparentNavBar)
         let customNavigationItem = UINavigationItem(title: "")
-        customNavigationBar.setItems([customNavigationItem], animated: false)
-        view.addSubview(customNavigationBar)
-
-        customNavigationBar.backgroundColor = UIColor.clearColor()
-        customNavigationBar.translucent = true
-        customNavigationBar.shadowImage = UIImage()
-        customNavigationBar.barStyle = UIBarStyle.BlackTranslucent
-        customNavigationBar.setBackgroundImage(UIImage(), forBarMetrics: UIBarMetrics.Default)
-
-        // add setting barItem
+        transparentNavBar.setItems([customNavigationItem], animated: false)
         let settingBarItem = UIBarButtonItem(image: UIImage(named: "setting_nav_item"), style: .Plain, target: self, action: "settingButtonTapped:")
         settingBarItem.tintColor = UIColor.whiteColor()
         customNavigationItem.rightBarButtonItem = settingBarItem
